Convert to ES6 classes, and modularise
