+++ .babelrc
@@ -1,11 +1,10 @@
 {
   "presets": [
     [
-      "env",
-      {
+      "@babel/preset-env", {
         "targets": {
-          "node": "6.10"
-        }
+            "node": "4"
+          }
       }
     ]
   ]
