+++ local.json
@@ -2,8 +2,8 @@
   "devtools_port": 9090,
   "startup_app": {
     "uuid": "docking-manager-example",
-    "name": "Docking Manager (v9+)",
-    "url": "http://localhost:8888/index-es6.html",
+    "name": "Docking Manager (v8)",
+    "url": "http://localhost:8888/index.html",
     "mainWindowOptions": {
       "taskbarIcon": "http://localhost:8888/openfin.ico",
       "autoShow": true,
@@ -15,7 +15,7 @@
   },
   "runtime": {
     "arguments": "--force-device-scale-factor=1",
-    "version": "9.61.29.24"
+    "version": "8.56.29.34"
   },
   "shortcut": {
     "company": "OpenFin",
