+++ package.json
@@ -1,12 +1,15 @@
 {
   "name": "snap-and-dock",
-  "version": "0.0.0",
+  "version": "2.0.0",
   "description": "a small framework for snap-to and docking of openfin windows",
   "main": "Docking.js",
   "scripts": {
     "build": "rollup Docking.js --o DockingManager-rolledup.js --f iife",
-    "test": "mocha --require babel-core/register --require babel-polyfill ./test/testSetup.js ./test/**/*-test.js",
-    "start": "http-server -c-1 -p 8888"
+    "test": "mocha --require @babel/register --require @babel/polyfill ./test/testSetup.js ./test/**/*-test.js",
+    "start": "http-server -c-1 -p 8888",
+    "launch": "npm run build && openfin -c local.json -l",
+    "launch-es6": "openfin -c local-es6.json -l",
+    "lint": "eslint Docking.js lib test"
   },
   "repository": {
     "type": "git",
@@ -19,13 +22,15 @@
   },
   "homepage": "https://github.com/openfin/snap-and-dock",
   "devDependencies": {
-    "babel-cli": "6.26.0",
-    "babel-core": "6.26.0",
-    "babel-polyfill": "6.26.0",
-    "babel-preset-env": "1.6.1",
+    "@babel/cli": "^7.0.0-beta.40",
+    "@babel/core": "^7.0.0-beta.40",
+    "@babel/polyfill": "^7.0.0-beta.40",
+    "@babel/preset-env": "^7.0.0-beta.40",
+    "@babel/register": "^7.0.0-beta.40",
     "eslint": "^4.18.2",
     "http-server": "^0.11.1",
-    "mocha": "5.0.1",
-    "rollup": "0.56.3"
+    "mocha": "^5.0.4",
+    "openfin-cli": "^1.1.5",
+    "rollup": "^0.56.5"
   }
 }
